‘リアルチャット機能’

+++ backend/Bookma/Bookma/routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-
+use App\Events\ChatPusher;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -61,12 +61,14 @@ Route::group(['middleware' => ['auth']], function () {
 
     //メッセージ一覧
     Route::get('/myPage/messagesList', 'MypageController@messages')->name('messagesList');
+    // リアルチャット
+    Route::get('/ChatRoom/user/{id}/json', 'ChatController@roomJson')->name('chat.room.json');
     // チャットルーム
     Route::get('/ChatRoom/user/{id}', 'ChatController@room')->name('chat.room');
     //チャット送受信
     Route::post('/ChatRoom/user/{id}/store', 'ChatController@store')->name('chat.store');
     //チャット削除
-    Route::delete('/ChatRoom/user/{id}/destroy', 'ChatController@destroy')->name('chat.destroy');
+    Route::delete('/ChatRoom/user/{id}/destroy/{userId}', 'ChatController@destroy')->name('chat.destroy');
 
     //お届け先の住所一覧
     Route::get('/myPage/shippingAddressList', 'MypageController@shippingAddressList')->name('shippingAddressList');
